Partial work towards #30 (real data tests).
Running the script run_test.sh in the pipeline_test directory will run the
species separation pipeline on a small sample data set using a number (which
can be added to) of different parameter settings; the resulting number of
hits/reads assigned to each species is then checked against the expected
amount. Without the -s option, only the species separation part of the pipeline
is run (this will hopefully allow us to run this part of the test on the Travis
continuous integration server). Specifying -s will run the whole pipeline,
including STAR mapping, which is more appropriate for local testing.