Fix percentage options for mismatch and minmatch.
The --mismatch-threshold option is the maximum percentage of *total* bases
allowed to be mismatches against the genome during filtering - for single-end
reads, the total number of bases is just the read length, whereas for
paired-end reads it is twice the read length. We check in this way because the
number of mismatches in STAR alignments reported for each hit is for the read
pair as a whole.

On the other hand, for the --minmatch-threshold option, that is, the maximum
percentage of read length allowed to be clipped during filtering, read length
refers to the length of a single read in both single- and paired-end cases.
This is because CIGAR strings are reported by STAR, and can thus be checked, on
a per-hit basis.